Style issue with v1.44 and media query for mobile
Version 1.44 has the width of .cookie-notice-container set to 100%. This results in this div being larger than the screen size in some cases. Adding box-sizing resolves this. 

On mobile/tablet, having the buttons appear across two lines isn't aesthetically pleasing when one button is to the right of the text. Adding a media query to force the buttons onto a new line resolves this.